pyproject.toml
[project]
name = "pdfrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.3",
    "inngest>=0.4.22",
    "llama-index-core>=0.14.13",
    "llama-index-readers-file>=0.5.6",
    "openai>=2.15.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.16.1",
    "streamlit>=1.50.0",
    "uvicorn>=0.39.0",
    "voyageai>=0.3.1",
]
